fix: настройки окна перенесены в натив
