+++ src/commonMain/kotlin/core/dsl/configuration/ConfigurationBuilder.kt
@@ -4,7 +4,6 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import core.configuration.Configuration
 import core.configuration.ConfigurationImpl
-import core.dsl.configuration.window.WindowBuilder
 
 /**
  * Builder конфигурации
@@ -28,24 +27,7 @@ class ConfigurationBuilder : Configuration() {
         this.logo = init
     }
 
-    /**
-     * Установка настроек окна (иконки и названия)
-     *
-     * ```kotlin
-     * val painter = painterResource("Good Tick.svg")
-     * config {
-     *      window {
-     *          icon = painter
-     *          title = "Simple title"
-     *      }
-     * }
-     * ```
-     */
-    fun window(init: WindowBuilder.() -> Unit){
-        val builder = WindowBuilder()
-        builder.init()
-        this.window = builder.build()
-    }
+
 
     /**
      * Создание [ConfigurationImpl] на основе полученных данных
@@ -55,7 +37,6 @@ class ConfigurationBuilder : Configuration() {
     internal fun build(): ConfigurationImpl {
         val configImpl = ConfigurationImpl()
         configImpl.logo = logo
-        configImpl.window = window
         return configImpl
     }
 }
