--- src/commonMain/kotlin/core/dsl/configuration/window/WindowBuilder.kt
@@ -1,8 +1,0 @@
-package core.dsl.configuration.window
-
-/**
- * Мультплатформенный WindowBuilder, для сборки окна
- *
- * @author Панков Вася (pank-su)
- */
-class WindowBuilder: WindowBuilderPlatform()
