+++ src/desktopMain/kotlin/core/extensions/BuildAll.kt
@@ -1,0 +1,16 @@
+package core.extensions
+
+import core.configuration.ConfigurationImpl
+import core.dsl.configuration.ConfigurationBuilder
+
+/**
+ * Добавление необходимых компонентов при сборке
+ *
+ * @author Панков Вася (pank-su)
+ */
+fun ConfigurationBuilder.build(): ConfigurationImpl {
+    val configuration = this.build()
+    configuration.keyMap = this.keyMap
+    configuration.window = this.window
+    return configuration
+}
